// src/app/shared/models/est_spk_kendaraan.model.ts
export class EstSpkKendaraan {
	id: number;
  lokasi_id: number;
  traksi_id: number;
  tanggal: string;
  no_spk: string;
  kontraktor_id: number;
  kendaraan_id: number;
  tanggal_mulai: string;
  tanggal_akhir: string;
  harga_sewa: number;
  harga_mob: number;
  uom_id: number;
  deskripsi: string;
}